+++ src/test/java/com/towermarsh/opendata/plugin/PluginSelectionResolverTest.java
@@ -64,6 +64,7 @@ class PluginSelectionResolverTest {
                 Optional.empty(),
                 OptionalInt.empty(),
                 false,
+                true,
                 false,
                 false,
                 false,